+++ types-override/openapi.d.ts
@@ -1,0 +1,18 @@
+export {default} from './lib/openapi.js';
+
+export type {
+  GenerateOpenAPIOptions,
+  OpenAPIDocument,
+  OpenAPIInfo,
+  OpenAPIContact,
+  OpenAPILicense,
+  OpenAPIServer,
+  OpenAPIServerVariable,
+  OpenAPIParameter,
+  OpenAPIMediaType,
+  OpenAPIRequestBody,
+  OpenAPIOperation,
+  OpenAPIPathItem,
+  OpenAPISecurityScheme,
+  OpenAPIComponents,
+} from './ergo-router.js';